feat: add guard and decorators

+++ src/home/home.controller.ts
@@ -8,10 +8,18 @@ import {
   Patch,
   Post,
   Query,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { HomeService } from './home.service';
-import { CreateHomeDto, HomeResponseDto, UpdateHomeDto } from './home.dto';
-import { PropertyType } from '@prisma/client';
+import {
+  CreateHomeDto,
+  HomeResponseDto,
+  InquireDto,
+  UpdateHomeDto,
+} from './home.dto';
+import { PropertyType, UserType } from '@prisma/client';
+import { User, UserInfo } from 'src/user/decorators/user.decorator';
+import { Roles } from 'src/user/decorators/roles.decorator';
 
 @Controller('home')
 export class HomeController {
@@ -47,21 +55,56 @@ export class HomeController {
     return await this.homeService.getHomeById(id);
   }
 
+  @Roles(UserType.REALTOR)
   @Post()
-  async createHome(@Body() body: CreateHomeDto): Promise<HomeResponseDto> {
-    return await this.homeService.createHome(body);
+  async createHome(
+    @Body() body: CreateHomeDto,
+    @User() user: UserInfo,
+  ): Promise<HomeResponseDto> {
+    return await this.homeService.createHome(body, user.id);
   }
 
+  @Roles(UserType.REALTOR)
   @Patch(':id')
   async updateHome(
     @Param('id', ParseIntPipe) id: number,
     @Body() body: UpdateHomeDto,
+    @User() user: UserInfo,
   ): Promise<HomeResponseDto> {
+    const realtor = await this.homeService.getRealtorByHomeId(user.id);
+    if (realtor.id !== user.id) throw new UnauthorizedException();
     return await this.homeService.updateHomeById(id, body);
   }
 
+  @Roles(UserType.REALTOR)
   @Delete(':id')
-  async deleteHome(@Param('id', ParseIntPipe) id: number): Promise<void> {
+  async deleteHome(
+    @Param('id', ParseIntPipe) id: number,
+    @User() user: UserInfo,
+  ): Promise<void> {
+    const realtor = await this.homeService.getRealtorByHomeId(user.id);
+    if (realtor.id !== user.id) throw new UnauthorizedException();
     return await this.homeService.deleteHomeById(id);
+  }
+
+  @Roles(UserType.BUYER)
+  @Post(':id/inquire')
+  async inquire(
+    @Param('id', ParseIntPipe) homeId: number,
+    @User() user: UserInfo,
+    @Body() { message }: InquireDto,
+  ) {
+    return await this.homeService.inquire(user, homeId, message);
+  }
+
+  @Roles(UserType.REALTOR)
+  @Get(':id/messages')
+  async getHomeMessages(
+    @Param('id', ParseIntPipe) id: number,
+    @User() user: UserInfo,
+  ) {
+    const realtor = await this.homeService.getRealtorByHomeId(id);
+    if (realtor.id !== user.id) throw new UnauthorizedException();
+    return await this.homeService.getMessagesByHome(id);
   }
 }
